Move compliance pluging to v2 system.
