--- lib/bundles/inspec-compliance/.kitchen.yml
@@ -1,20 +1,0 @@
----
-driver:
-  name: vagrant
-  synced_folders:
-    - ['../../../', '/inspec']
-  network:
-    - ['private_network', {ip: '192.168.251.2'}]
-
-provisioner:
-  name: shell
-
-verifier:
-  name: inspec
-
-platforms:
-  - name: ubuntu-14.04
-suites:
-  - name: default
-    run_list:
-    attributes:
