+++ lib/plugins/inspec-compliance/lib/inspec-compliance/images/cc-token.png
