+++ lib/plugins/inspec-compliance/test/integration/default/cli.rb
