+++ lib/plugins/inspec-compliance/test/unit/api/login_test.rb
@@ -1,6 +1,9 @@
-require 'helper'
+require 'minitest/autorun'
+require 'mocha/setup'
+require 'webmock/minitest'
+require_relative '../../../lib/inspec-compliance/api.rb'
 
-describe Compliance::API do
+describe InspecPlugins::Compliance::API do
   let(:automate_options) do
     {
       'server' => 'https://automate.example.com',
@@ -49,13 +52,13 @@ describe Compliance::API do
   describe '.login' do
     describe 'when target is a Chef Automate2 server' do
       before do
-        Compliance::API.expects(:determine_server_type).returns(:automate2)
+        InspecPlugins::Compliance::API.expects(:determine_server_type).returns(:automate2)
       end
 
       it 'raises an error if `--user` is missing' do
         options = automate_options
         options.delete('user')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify a user.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -64,7 +67,7 @@ describe Compliance::API do
         options = automate_options
         options.delete('token')
         options.delete('dctoken')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify a token.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -73,9 +76,9 @@ describe Compliance::API do
         stub_request(:get, automate_options['server'] + '/compliance/version')
           .to_return(status: 200, body: '', headers: {})
         options = automate_options
-        Compliance::Configuration.expects(:new).returns(fake_config)
+        InspecPlugins::Compliance::Configuration.expects(:new).returns(fake_config)
 
-        Compliance::API.login(options)
+        InspecPlugins::Compliance::API.login(options)
         fake_config['automate']['ent'].must_equal('automate')
         fake_config['automate']['token_type'].must_equal('dctoken')
         fake_config['user'].must_equal('someone')
@@ -87,13 +90,13 @@ describe Compliance::API do
 
     describe 'when target is a Chef Automate server' do
       before do
-        Compliance::API.expects(:determine_server_type).returns(:automate)
+        InspecPlugins::Compliance::API.expects(:determine_server_type).returns(:automate)
       end
 
       it 'raises an error if `--user` is missing' do
         options = automate_options
         options.delete('user')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify a user.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -101,7 +104,7 @@ describe Compliance::API do
       it 'raises an error if `--ent` is missing' do
         options = automate_options
         options.delete('ent')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify an enterprise.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -110,7 +113,7 @@ describe Compliance::API do
         options = automate_options
         options.delete('token')
         options.delete('dctoken')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify a token.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -119,9 +122,9 @@ describe Compliance::API do
         stub_request(:get, automate_options['server'] + '/compliance/version')
           .to_return(status: 200, body: '', headers: {})
         options = automate_options
-        Compliance::Configuration.expects(:new).returns(fake_config)
+        InspecPlugins::Compliance::Configuration.expects(:new).returns(fake_config)
 
-        Compliance::API.login(options)
+        InspecPlugins::Compliance::API.login(options)
         fake_config['automate']['ent'].must_equal('automate')
         fake_config['automate']['token_type'].must_equal('usertoken')
         fake_config['user'].must_equal('someone')
@@ -133,14 +136,14 @@ describe Compliance::API do
 
     describe 'when target is a Chef Compliance server' do
       before do
-        Compliance::API.expects(:determine_server_type).returns(:compliance)
+        InspecPlugins::Compliance::API.expects(:determine_server_type).returns(:compliance)
       end
 
       it 'raises an error if `--user` and `--refresh-token` are missing' do
         options = automate_options
         options.delete('user')
         options.delete('refresh_token')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify a.*--user.*--refresh-token.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -150,7 +153,7 @@ describe Compliance::API do
         options.delete('password')
         options.delete('token')
         options.delete('refresh_token')
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify.*--password.*--token.*--refresh-token.*/)
         err.message.lines.length.must_equal(1)
       end
@@ -159,9 +162,9 @@ describe Compliance::API do
         stub_request(:get, compliance_options['server'] + '/api/version')
           .to_return(status: 200, body: '', headers: {})
         options = compliance_options
-        Compliance::Configuration.expects(:new).returns(fake_config)
+        InspecPlugins::Compliance::Configuration.expects(:new).returns(fake_config)
 
-        Compliance::API.login(options)
+        InspecPlugins::Compliance::API.login(options)
         fake_config['user'].must_equal('someone')
         fake_config['server'].must_equal('https://compliance.example.com/api')
         fake_config['server_type'].must_equal('compliance')
@@ -172,14 +175,14 @@ describe Compliance::API do
     describe 'when target is neither a Chef Compliance nor Chef Automate server' do
       it 'raises an error if `https://SERVER` is missing' do
         options = {}
-        err = proc { Compliance::API.login(options) }.must_raise(ArgumentError)
+        err = proc { InspecPlugins::Compliance::API.login(options) }.must_raise(ArgumentError)
         err.message.must_match(/Please specify a server.*/)
         err.message.lines.length.must_equal(1)
       end
 
       it 'rasies a `CannotDetermineServerType` error' do
-        Compliance::API.expects(:determine_server_type).returns(nil)
-        err = proc { Compliance::API.login(automate_options) }.must_raise(StandardError)
+        InspecPlugins::Compliance::API.expects(:determine_server_type).returns(nil)
+        err = proc { InspecPlugins::Compliance::API.login(automate_options) }.must_raise(StandardError)
         err.message.must_match(/Unable to determine/)
       end
     end
