+++ lib/plugins/inspec-compliance/test/unit/api_test.rb
@@ -1,6 +1,8 @@
-require 'helper'
+require 'minitest/autorun'
+require 'mocha/setup'
+require_relative '../../lib/inspec-compliance/api.rb'
 
-describe Compliance::API do
+describe InspecPlugins::Compliance::API do
   let(:profiles_response) do
     [{ 'name'=>'apache-baseline',
       'title'=>'DevSec Apache Baseline',
@@ -49,15 +51,15 @@ describe Compliance::API do
     end
 
     before do
-      Compliance::API.expects(:get_headers).returns(headers)
+      InspecPlugins::Compliance::API.expects(:get_headers).returns(headers)
     end
 
     describe 'when a 404 is received' do
       it 'should return an empty hash' do
         response = mock
         response.stubs(:code).returns('404')
-        Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
-        Compliance::API.version(config).must_equal({})
+        InspecPlugins::Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
+        InspecPlugins::Compliance::API.version(config).must_equal({})
       end
     end
 
@@ -66,8 +68,8 @@ describe Compliance::API do
         response = mock
         response.stubs(:code).returns('200')
         response.stubs(:body).returns(nil)
-        Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
-        Compliance::API.version(config).must_equal({})
+        InspecPlugins::Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
+        InspecPlugins::Compliance::API.version(config).must_equal({})
       end
     end
 
@@ -76,8 +78,8 @@ describe Compliance::API do
         response = mock
         response.stubs(:code).returns('200')
         response.stubs(:body).returns('')
-        Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
-        Compliance::API.version(config).must_equal({})
+        InspecPlugins::Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
+        InspecPlugins::Compliance::API.version(config).must_equal({})
       end
     end
 
@@ -86,8 +88,8 @@ describe Compliance::API do
         response = mock
         response.stubs(:code).returns('200')
         response.stubs(:body).returns('{"api":"compliance"}')
-        Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
-        Compliance::API.version(config).must_equal({})
+        InspecPlugins::Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
+        InspecPlugins::Compliance::API.version(config).must_equal({})
       end
     end
 
@@ -96,8 +98,8 @@ describe Compliance::API do
         response = mock
         response.stubs(:code).returns('200')
         response.stubs(:body).returns('{"api":"compliance","version":""}')
-        Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
-        Compliance::API.version(config).must_equal({})
+        InspecPlugins::Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
+        InspecPlugins::Compliance::API.version(config).must_equal({})
       end
     end
 
@@ -106,8 +108,8 @@ describe Compliance::API do
         response = mock
         response.stubs(:code).returns('200')
         response.stubs(:body).returns('{"api":"compliance","version":"1.2.3"}')
-        Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
-        Compliance::API.version(config).must_equal({ 'version' => '1.2.3', 'api' => 'compliance' })
+        InspecPlugins::Compliance::HTTP.expects(:get).with('myserver/version', 'test-headers', true).returns(response)
+        InspecPlugins::Compliance::API.version(config).must_equal({ 'version' => '1.2.3', 'api' => 'compliance' })
       end
     end
   end
@@ -115,80 +117,80 @@ describe Compliance::API do
   describe 'automate/compliance is? checks' do
     describe 'when the config has a compliance server_type' do
       it 'automate/compliance server is? methods return correctly' do
-        config = Compliance::Configuration.new
+        config = InspecPlugins::Compliance::Configuration.new
         config.clean
         config['server_type'] = 'compliance'
-        Compliance::API.is_compliance_server?(config).must_equal true
-        Compliance::API.is_automate_server?(config).must_equal false
-        Compliance::API.is_automate_server_pre_080?(config).must_equal false
-        Compliance::API.is_automate_server_080_and_later?(config).must_equal false
-        Compliance::API.is_automate2_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_compliance_server?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server_pre_080?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server_080_and_later?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate2_server?(config).must_equal false
       end
     end
 
     describe 'when the config has a automate2 server_type' do
       it 'automate/compliance server is? methods return correctly' do
-        config = Compliance::Configuration.new
+        config = InspecPlugins::Compliance::Configuration.new
         config.clean
         config['server_type'] = 'automate2'
-        Compliance::API.is_compliance_server?(config).must_equal false
-        Compliance::API.is_automate_server?(config).must_equal false
-        Compliance::API.is_automate_server_pre_080?(config).must_equal false
-        Compliance::API.is_automate_server_080_and_later?(config).must_equal false
-        Compliance::API.is_automate2_server?(config).must_equal true
+        InspecPlugins::Compliance::API.is_compliance_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server_pre_080?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server_080_and_later?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate2_server?(config).must_equal true
       end
     end
 
     describe 'when the config has an automate server_type and no version key' do
       it 'automate/compliance server is? methods return correctly' do
-        config = Compliance::Configuration.new
+        config = InspecPlugins::Compliance::Configuration.new
         config.clean
         config['server_type'] = 'automate'
-        Compliance::API.is_compliance_server?(config).must_equal false
-        Compliance::API.is_automate_server?(config).must_equal true
-        Compliance::API.is_automate_server_pre_080?(config).must_equal true
-        Compliance::API.is_automate_server_080_and_later?(config).must_equal false
-        Compliance::API.is_automate2_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_compliance_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_pre_080?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_080_and_later?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate2_server?(config).must_equal false
       end
     end
 
     describe 'when the config has an automate server_type and a version key that is not a hash' do
       it 'automate/compliance server is? methods return correctly' do
-        config = Compliance::Configuration.new
+        config = InspecPlugins::Compliance::Configuration.new
         config.clean
         config['server_type'] = 'automate'
         config['version'] = '1.2.3'
-        Compliance::API.is_compliance_server?(config).must_equal false
-        Compliance::API.is_automate_server?(config).must_equal true
-        Compliance::API.is_automate_server_pre_080?(config).must_equal true
-        Compliance::API.is_automate_server_080_and_later?(config).must_equal false
-        Compliance::API.is_automate2_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_compliance_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_pre_080?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_080_and_later?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate2_server?(config).must_equal false
       end
     end
 
     describe 'when the config has an automate server_type and a version hash with no version' do
       it 'automate/compliance server is? methods return correctly' do
-        config = Compliance::Configuration.new
+        config = InspecPlugins::Compliance::Configuration.new
         config.clean
         config['server_type'] = 'automate'
         config['version'] = {}
-        Compliance::API.is_compliance_server?(config).must_equal false
-        Compliance::API.is_automate_server?(config).must_equal true
-        Compliance::API.is_automate_server_pre_080?(config).must_equal true
-        Compliance::API.is_automate_server_080_and_later?(config).must_equal false
+        InspecPlugins::Compliance::API.is_compliance_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_pre_080?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_080_and_later?(config).must_equal false
       end
     end
 
     describe 'when the config has an automate server_type and a version hash with a version' do
       it 'automate/compliance server is? methods return correctly' do
-        config = Compliance::Configuration.new
+        config = InspecPlugins::Compliance::Configuration.new
         config.clean
         config['server_type'] = 'automate'
         config['version'] = { 'version' => '0.8.1' }
-        Compliance::API.is_compliance_server?(config).must_equal false
-        Compliance::API.is_automate_server?(config).must_equal true
-        Compliance::API.is_automate_server_pre_080?(config).must_equal false
-        Compliance::API.is_automate_server_080_and_later?(config).must_equal true
+        InspecPlugins::Compliance::API.is_compliance_server?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server?(config).must_equal true
+        InspecPlugins::Compliance::API.is_automate_server_pre_080?(config).must_equal false
+        InspecPlugins::Compliance::API.is_automate_server_080_and_later?(config).must_equal true
       end
     end
   end
@@ -196,54 +198,54 @@ describe Compliance::API do
   describe '.server_version_from_config' do
     it 'returns nil when the config has no version key' do
       config = {}
-      Compliance::API.server_version_from_config(config).must_be_nil
+      InspecPlugins::Compliance::API.server_version_from_config(config).must_be_nil
     end
 
     it 'returns nil when the version value is not a hash' do
       config = { 'version' => '123' }
-      Compliance::API.server_version_from_config(config).must_be_nil
+      InspecPlugins::Compliance::API.server_version_from_config(config).must_be_nil
     end
 
     it 'returns nil when the version value is a hash but has no version key inside' do
       config = { 'version' => {} }
-      Compliance::API.server_version_from_config(config).must_be_nil
+      InspecPlugins::Compliance::API.server_version_from_config(config).must_be_nil
     end
 
     it 'returns the version if the version value is a hash containing a version' do
       config = { 'version' => { 'version' => '1.2.3' } }
-      Compliance::API.server_version_from_config(config).must_equal '1.2.3'
+      InspecPlugins::Compliance::API.server_version_from_config(config).must_equal '1.2.3'
     end
   end
 
   describe 'profile_split' do
     it 'handles a profile without version' do
-      Compliance::API.profile_split('admin/apache-baseline').must_equal ['admin', 'apache-baseline', nil]
+      InspecPlugins::Compliance::API.profile_split('admin/apache-baseline').must_equal ['admin', 'apache-baseline', nil]
     end
 
     it 'handles a profile with a version' do
-      Compliance::API.profile_split('admin/apache-baseline#2.0.1').must_equal ['admin', 'apache-baseline', '2.0.1']
+      InspecPlugins::Compliance::API.profile_split('admin/apache-baseline#2.0.1').must_equal ['admin', 'apache-baseline', '2.0.1']
     end
   end
 
   describe 'target_url' do
     it 'handles a automate profile with and without version' do
-      config = Compliance::Configuration.new
+      config = InspecPlugins::Compliance::Configuration.new
       config.clean
       config['server_type'] = 'automate'
       config['server'] = 'https://myautomate'
       config['version'] = '1.6.99'
-      Compliance::API.target_url(config, 'admin/apache-baseline').must_equal       'https://myautomate/profiles/admin/apache-baseline/tar'
-      Compliance::API.target_url(config, 'admin/apache-baseline#2.0.2').must_equal 'https://myautomate/profiles/admin/apache-baseline/version/2.0.2/tar'
+      InspecPlugins::Compliance::API.target_url(config, 'admin/apache-baseline').must_equal       'https://myautomate/profiles/admin/apache-baseline/tar'
+      InspecPlugins::Compliance::API.target_url(config, 'admin/apache-baseline#2.0.2').must_equal 'https://myautomate/profiles/admin/apache-baseline/version/2.0.2/tar'
     end
 
     it 'handles a chef-compliance profile with and without version' do
-      config = Compliance::Configuration.new
+      config = InspecPlugins::Compliance::Configuration.new
       config.clean
       config['server_type'] = 'compliance'
       config['server'] = 'https://mychefcompliance'
       config['version'] = '1.1.2'
-      Compliance::API.target_url(config, 'admin/apache-baseline').must_equal       'https://mychefcompliance/owners/admin/compliance/apache-baseline/tar'
-      Compliance::API.target_url(config, 'admin/apache-baseline#2.0.2').must_equal 'https://mychefcompliance/owners/admin/compliance/apache-baseline/tar'
+      InspecPlugins::Compliance::API.target_url(config, 'admin/apache-baseline').must_equal       'https://mychefcompliance/owners/admin/compliance/apache-baseline/tar'
+      InspecPlugins::Compliance::API.target_url(config, 'admin/apache-baseline#2.0.2').must_equal 'https://mychefcompliance/owners/admin/compliance/apache-baseline/tar'
     end
   end
 
@@ -253,7 +255,7 @@ describe Compliance::API do
       # skipping for that specific version
       return if RUBY_VERSION = '2.3.3'
 
-      config = Compliance::Configuration.new
+      config = InspecPlugins::Compliance::Configuration.new
       config.clean
       config['owner'] = 'admin'
       config['server_type'] = 'automate'
@@ -266,10 +268,10 @@ describe Compliance::API do
         .with(headers: { 'Accept'=>'*/*', 'Accept-Encoding'=>'gzip;q=1.0,deflate;q=0.6,identity;q=0.3', 'Chef-Delivery-Enterprise'=>'automate', 'User-Agent'=>'Ruby', 'X-Data-Collector-Token'=>'' })
         .to_return(status: 200, body: profiles_response.to_json, headers: {})
 
-      Compliance::API.exist?(config, 'admin/apache-baseline').must_equal true
-      Compliance::API.exist?(config, 'admin/apache-baseline#2.0.1').must_equal true
-      Compliance::API.exist?(config, 'admin/apache-baseline#2.0.999').must_equal false
-      Compliance::API.exist?(config, 'admin/missing-in-action').must_equal false
+      InspecPlugins::Compliance::API.exist?(config, 'admin/apache-baseline').must_equal true
+      InspecPlugins::Compliance::API.exist?(config, 'admin/apache-baseline#2.0.1').must_equal true
+      InspecPlugins::Compliance::API.exist?(config, 'admin/apache-baseline#2.0.999').must_equal false
+      InspecPlugins::Compliance::API.exist?(config, 'admin/missing-in-action').must_equal false
     end
   end
 
@@ -288,25 +290,25 @@ describe Compliance::API do
     it 'returns `:automate2` when a 400 is received from `https://URL/dex/auth`' do
       good_response.stubs(:code).returns('400')
 
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate2_endpoint, headers, insecure)
                       .returns(good_response)
 
-      Compliance::API.determine_server_type(url, insecure).must_equal(:automate2)
+      InspecPlugins::Compliance::API.determine_server_type(url, insecure).must_equal(:automate2)
     end
 
     it 'returns `:automate` when a 401 is received from `https://URL/compliance/version`' do
       good_response.stubs(:code).returns('401')
       bad_response.stubs(:code).returns('404')
 
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate2_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate_endpoint, headers, insecure)
                       .returns(good_response)
 
-      Compliance::API.determine_server_type(url, insecure).must_equal(:automate)
+      InspecPlugins::Compliance::API.determine_server_type(url, insecure).must_equal(:automate)
     end
 
     # Chef Automate currently returns 401 for `/compliance/version` but some
@@ -317,67 +319,67 @@ describe Compliance::API do
       good_response.stubs(:code).returns('200')
       good_response.stubs(:body).returns('Are You Looking For the Chef Server?')
 
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate2_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate_endpoint, headers, insecure)
                       .returns(good_response)
 
-      Compliance::API.determine_server_type(url, insecure).must_equal(:automate)
+      InspecPlugins::Compliance::API.determine_server_type(url, insecure).must_equal(:automate)
     end
 
     it 'returns `nil` if a 200 is received from `https://URL/compliance/version` but not redirected to Chef Manage' do
       bad_response.stubs(:code).returns('200')
       bad_response.stubs(:body).returns('No Chef Manage here')
 
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate2_endpoint, headers, insecure)
                       .returns(bad_response)
 
       mock_compliance_response = mock
       mock_compliance_response.stubs(:code).returns('404')
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + compliance_endpoint, headers, insecure)
                       .returns(mock_compliance_response)
 
-      Compliance::API.determine_server_type(url, insecure).must_be_nil
+      InspecPlugins::Compliance::API.determine_server_type(url, insecure).must_be_nil
     end
 
     it 'returns `:compliance` when a 200 is received from `https://URL/api/version`' do
       good_response.stubs(:code).returns('200')
       bad_response.stubs(:code).returns('404')
 
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate2_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + compliance_endpoint, headers, insecure)
                       .returns(good_response)
 
-      Compliance::API.determine_server_type(url, insecure).must_equal(:compliance)
+      InspecPlugins::Compliance::API.determine_server_type(url, insecure).must_equal(:compliance)
     end
 
     it 'returns `nil` if it cannot determine the server type' do
       bad_response.stubs(:code).returns('404')
 
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate2_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + automate_endpoint, headers, insecure)
                       .returns(bad_response)
-      Compliance::HTTP.expects(:get)
+      InspecPlugins::Compliance::HTTP.expects(:get)
                       .with(url + compliance_endpoint, headers, insecure)
                       .returns(bad_response)
 
-      Compliance::API.determine_server_type(url, insecure).must_be_nil
+      InspecPlugins::Compliance::API.determine_server_type(url, insecure).must_be_nil
     end
   end
 end
