+++ lib/plugins/inspec-compliance/test/unit/target_test.rb
@@ -1,10 +1,12 @@
-require 'helper'
+require 'minitest/autorun'
+require 'mocha/setup'
+require_relative '../../lib/inspec-compliance/api.rb'
 
-describe Compliance::Fetcher do
+describe InspecPlugins::Compliance::Fetcher do
   let(:config) { { 'server' => 'myserver' } }
 
   describe 'the check_compliance_token method' do
-    let(:fetcher) { fetcher = Compliance::Fetcher.new('a/bad/url', config) }
+    let(:fetcher) { fetcher = InspecPlugins::Compliance::Fetcher.new('a/bad/url', config) }
 
     it 'returns without error if token is set' do
       config['token'] = 'my-token'
@@ -18,59 +20,59 @@ describe Compliance::Fetcher do
   end
 
   describe 'when the server is an automate2 server' do
-    before { Compliance::API.expects(:is_automate2_server?).with(config).returns(true) }
+    before { InspecPlugins::Compliance::API.expects(:is_automate2_server?).with(config).returns(true) }
 
     it 'returns the correct owner and profile name' do
       config['profile'] = ['admin', 'ssh-baseline', nil]
-      fetcher = Compliance::Fetcher.new('myserver/profile', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('myserver/profile', config)
       fetcher.send(:compliance_profile_name).must_equal 'admin/ssh-baseline'
     end
   end
 
   describe 'when the server is an automate server pre-0.8.0' do
-    before { Compliance::API.expects(:is_automate_server_pre_080?).with(config).returns(true) }
+    before { InspecPlugins::Compliance::API.expects(:is_automate_server_pre_080?).with(config).returns(true) }
 
     it 'returns the correct profile name when the url is correct' do
-      fetcher = Compliance::Fetcher.new('myserver/myowner/myprofile/tar', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('myserver/myowner/myprofile/tar', config)
       fetcher.send(:compliance_profile_name).must_equal 'myowner/myprofile'
     end
 
     it 'raises an exception if the url is malformed' do
-      fetcher = Compliance::Fetcher.new('a/bad/url', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('a/bad/url', config)
       proc { fetcher.send(:compliance_profile_name) }.must_raise RuntimeError
     end
   end
 
   describe 'when the server is an automate server 0.8.0-or-later' do
     before do
-      Compliance::API.expects(:is_automate_server_pre_080?).with(config).returns(false)
-      Compliance::API.expects(:is_automate_server_080_and_later?).with(config).returns(true)
+      InspecPlugins::Compliance::API.expects(:is_automate_server_pre_080?).with(config).returns(false)
+      InspecPlugins::Compliance::API.expects(:is_automate_server_080_and_later?).with(config).returns(true)
     end
 
     it 'returns the correct profile name when the url is correct' do
-      fetcher = Compliance::Fetcher.new('myserver/profiles/myowner/myprofile/tar', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('myserver/profiles/myowner/myprofile/tar', config)
       fetcher.send(:compliance_profile_name).must_equal 'myowner/myprofile'
     end
 
     it 'raises an exception if the url is malformed' do
-      fetcher = Compliance::Fetcher.new('a/bad/url', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('a/bad/url', config)
       proc { fetcher.send(:compliance_profile_name) }.must_raise RuntimeError
     end
   end
 
   describe 'when the server is not an automate server (likely a compliance server)' do
     before do
-      Compliance::API.expects(:is_automate_server_pre_080?).with(config).returns(false)
-      Compliance::API.expects(:is_automate_server_080_and_later?).with(config).returns(false)
+      InspecPlugins::Compliance::API.expects(:is_automate_server_pre_080?).with(config).returns(false)
+      InspecPlugins::Compliance::API.expects(:is_automate_server_080_and_later?).with(config).returns(false)
     end
 
     it 'returns the correct profile name when the url is correct' do
-      fetcher = Compliance::Fetcher.new('myserver/owners/myowner/compliance/myprofile/tar', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('myserver/owners/myowner/compliance/myprofile/tar', config)
       fetcher.send(:compliance_profile_name).must_equal 'myowner/myprofile'
     end
 
     it 'raises an exception if the url is malformed' do
-      fetcher = Compliance::Fetcher.new('a/bad/url', config)
+      fetcher = InspecPlugins::Compliance::Fetcher.new('a/bad/url', config)
       proc { fetcher.send(:compliance_profile_name) }.must_raise RuntimeError
     end
   end
@@ -95,24 +97,24 @@ describe Compliance::Fetcher do
           'latest_version'=>'' }]
     end
     before do
-      Compliance::Configuration.expects(:new).returns({ 'token' => '123abc', 'server' => 'https://a2.instance.com' })
+      InspecPlugins::Compliance::Configuration.expects(:new).returns({ 'token' => '123abc', 'server' => 'https://a2.instance.com' })
     end
 
     it 'returns the correct profile name when parsing url' do
-      Compliance::API.stubs(:profiles).returns(['success', profiles_result])
-      fetcher = Compliance::Fetcher.resolve('compliance://admin/ssh-baseline')
+      InspecPlugins::Compliance::API.stubs(:profiles).returns(['success', profiles_result])
+      fetcher = InspecPlugins::Compliance::Fetcher.resolve('compliance://admin/ssh-baseline')
       assert = ['admin', 'ssh-baseline', nil]
       fetcher.instance_variable_get(:"@config")['profile'].must_equal assert
     end
 
     it 'returns the correct profile name when parsing compliance hash' do
-      Compliance::API.stubs(:profiles).returns(['success', profiles_result])
+      InspecPlugins::Compliance::API.stubs(:profiles).returns(['success', profiles_result])
       hash = {
         target: 'https://a2.instance.com/api/v0/compliance/tar',
         compliance: 'admin/ssh-baseline',
         sha256: '132j1kjdasfasdoaefaewo12312',
       }
-      fetcher = Compliance::Fetcher.resolve(hash)
+      fetcher = InspecPlugins::Compliance::Fetcher.resolve(hash)
       assert = ['admin', 'ssh-baseline', nil]
       fetcher.instance_variable_get(:"@config")['profile'].must_equal assert
     end
@@ -139,14 +141,14 @@ describe Compliance::Fetcher do
     end
 
     before do
-      Compliance::Configuration.expects(:new).returns({ 'token' => '123abc', 'server' => 'https://a2.instance.com' })
+      InspecPlugins::Compliance::Configuration.expects(:new).returns({ 'token' => '123abc', 'server' => 'https://a2.instance.com' })
     end
 
     it 'contains the upstream_sha256' do
-      Compliance::API.stubs(:profiles).returns(['success', profiles_result])
+      InspecPlugins::Compliance::API.stubs(:profiles).returns(['success', profiles_result])
       prof = profiles_result[0]
       target = "compliance://#{prof['owner']}/#{prof['name']}"
-      fetcher = Compliance::Fetcher.resolve(target)
+      fetcher = InspecPlugins::Compliance::Fetcher.resolve(target)
       fetcher.upstream_sha256.must_equal prof['sha256']
     end
   end
