+++ test/helper.rb
@@ -37,7 +37,7 @@ require 'inspec/impact'
 require 'fetchers/mock'
 require 'inspec/dependencies/cache'
 
-require_relative '../lib/bundles/inspec-compliance'
+require_relative '../lib/bundles/inspec-supermarket'
 
 require 'train'
 CMD = Train.create('local', command_runner: :generic).connection
